Food Edit Dashboard

+++ frontend/src/views/Dashboard/Components/FoodsTab/FoodsTab.jsx
@@ -3,10 +3,9 @@ import { useNavigate } from 'react-router-dom';
 import axios from 'axios';
 import './FoodsTab.css';
 
-const FoodsTab = () => {
+const FoodsTab = ({ onEditFood }) => {
   const [foods, setFoods] = useState([]);
-  const navigate = useNavigate();
-  // Función para obtener los alimentos
+
   useEffect(() => {
     const fetchFoods = async () => {
       try {
@@ -21,11 +20,13 @@ const FoodsTab = () => {
 
   // Función para eliminar un alimento
   const deleteFood = async (foodId) => {
-    try {
-      await axios.delete(`http://localhost:3000/foods/${foodId}`);
-      setFoods((prevFoods) => prevFoods.filter((food) => food.id !== foodId));
-    } catch (error) {
-      console.error('Error deleting food:', error.message);
+    if (window.confirm('Are you sure you want to delete this food?')) {
+      try {
+        await axios.delete(`http://localhost:3000/foods/${foodId}`);
+        setFoods((prevFoods) => prevFoods.filter((food) => food.id !== foodId));
+      } catch (error) {
+        console.error('Error deleting food:', error.message);
+      }
     }
   };
 
@@ -61,8 +62,8 @@ const FoodsTab = () => {
               <td>{food.quanty}</td>
               <td>{food.unit}</td>
               <td>
-                
-                <button onClick={() => deleteFood(food.id)} className="delete-button">Delete</button>
+                <button onClick={() => onEditFood(food)}>Edit</button>
+                <button onClick={() => deleteFood(food.id)}>Delete</button>
               </td>
             </tr>
           ))}
